Minor drawing and build bug fixes.
* Fixed clipping bug when USE_WX_ZOOM is enabled that causes rats nest
  to be drawn incorrectly.
* Fix ambiguous function error in trigo.cpp when building with MSVS.
* Add instructions for building wxWidgets with graphics context
  enabled (GDI+) using MinGW/MSYS to COMPILING.txt.
* Initial EESchema find dialog work.
* Set modified flag when using PCBNew global deletion dialog to prevent
  closing without warning user of changes.

+++ common/trigo.cpp
@@ -402,9 +402,9 @@ wxPoint TwoPointVector( wxPoint startPoint, wxPoint endPoint )
 
 double DistanceLinePoint( wxPoint linePointA, wxPoint linePointB, wxPoint referencePoint )
 {
-    return fabs( (linePointB.x - linePointA.x) * (linePointA.y - referencePoint.y) -
-                (linePointA.x - referencePoint.x ) * (linePointB.y - linePointA.y) )
-           / EuclideanNorm( TwoPointVector( linePointA, linePointB ) );
+    return fabs( (double) ( (linePointB.x - linePointA.x) * (linePointA.y - referencePoint.y) -
+                 (linePointA.x - referencePoint.x ) * (linePointB.y - linePointA.y) )
+                  / EuclideanNorm( TwoPointVector( linePointA, linePointB ) ) );
 }
